Add support for the ipnetwork prefix types
In the Rust ecosystem, both the "ipnet" and "ipnetwork" crates are
widely used for representing IP prefixes. This crate, however,
implements the `Prefix` trait only for the "ipnet" prefix types.

In this commit, we introduce an optional feature that enables
the implementation of the `Prefix` trait for the `Ipv4Network` and
`Ipv6Network` types of the "ipnetwork" crate, making this crate more
accessible to a wider audience.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>